isolated ThemeContext from CurrencyContext
