+++ src/App.tsx
@@ -1,13 +1,16 @@
 import { AppRouter } from "./routes";
 import { RouterProvider } from "react-router-dom";
 import { CurrencyContextProvider } from "./contexts";
+import { ThemeContextProvider } from "./contexts/ThemeContext";
 
 function App() {
   return (
     <>
-      <CurrencyContextProvider>
-        <RouterProvider router={AppRouter} />
-      </CurrencyContextProvider>
+      <ThemeContextProvider>
+        <CurrencyContextProvider>
+          <RouterProvider router={AppRouter} />
+        </CurrencyContextProvider>
+      </ThemeContextProvider>
     </>
   );
 }
