+++ src/components/index.ts
@@ -4,4 +4,3 @@ export * from "./ImageWithFallback";
 export * from "./Loader";
 export * from "./CurrencySelect";
 export * from "./Card";
-export * from "./Slider";
